Changed Shibboleth_UP class
