+++ src/bootup/models/shibboleth/shibboleth.js
@@ -1,6 +1,3 @@
-const slug = require("slug");
-const path = require("path");
-
 const Pathfinder = global.Pathfinder;
 const isNull = require(Pathfinder.absPathInSrcFolder("/utils/null.js")).isNull;
 const Logger = require(Pathfinder.absPathInSrcFolder("utils/logger.js")).Logger;
@@ -8,7 +5,6 @@ const User = require(Pathfinder.absPathInSrcFolder("models/user.js")).User;
 const UploadManager = require(Pathfinder.absPathInSrcFolder("/models/uploads/upload_manager.js")).UploadManager;
 
 const Config = require(Pathfinder.absPathInSrcFolder("models/meta/config.js")).Config;
-const _ = require("underscore");
 const fs = require("fs");
 
 class Shibboleth
@@ -112,9 +108,74 @@ class Shibboleth
         }
     }
 
-    registerAuthenticationRoutes (app, passport)
+    registerAuthenticationRoutes(app, passport)
     {
+        const self = this;
+        let saml = require("passport-saml");
+        let samlStrategy = new saml.Strategy({
+            // URL that goes from the Identity Provider -> Service Provider
+            callbackUrl: self.getCallbackURl(),
+            // URL that goes from the Service Provider -> Identity Provider
+            entryPoint: self.getEntryPoint(),
+            // Usually specified as `/shibboleth` from site root
+            issuer: self.getIssuer(),
+            identifierFormat: null,
+            // Service Provider private key
+            decryptionPvk: self.getKey(),
+            // Service Provider Certificate
+            privateCert: self.getKey(),
+            // Identity Provider's public key
+            cert: self.getIdpCert(),
+            validateInResponseTo: false,
+            disableRequestedAuthnContext: true
+        }, function(profile, done) {
+            return done(null, profile);
+        });
 
+        passport.use(samlStrategy);
+
+        function ensureAuthenticated(req, res, next) {
+            if (req.isAuthenticated())
+                return next();
+            else
+                return res.redirect("/Shibboleth/login");
+        }
+
+        app.get("/Shibboleth",
+            ensureAuthenticated,
+            function(req, res) {
+                res.send('Authenticated');
+            }
+        );
+
+        app.get("/Shibboleth/login",
+            passport.authenticate("saml", { failureRedirect: "/Shibboleth/login/fail" }),
+            function (req, res) {
+                res.redirect("/");
+            }
+        );
+
+        app.post("/Shibboleth/login/callback",
+            passport.authenticate("saml", { failureRedirect: "/Shibboleth/login/fail" }),
+            function(req, res) {
+                Logger.log("info", "will check req.user!!");
+                self.auth(req, res);
+            }
+        );
+
+        app.get("/Shibboleth/login/fail",
+            function(req, res) {
+                Logger.log("Login failed!");
+                res.status(401).send("Login failed");
+            }
+        );
+
+        app.get("/Shibboleth.sso/Metadata",
+            function(req, res) {
+                res.type("application/xml");
+                res.status(200).send(samlStrategy.generateServiceProviderMetadata(self.getCert()));
+            }
+        );
     }
 
     loginUser(user, req, res)
